+++ pkg/sink/codec/canal/canal_json_txn_decoder.go
@@ -122,21 +122,21 @@ func (d *canalJSONTxnEventDecoder) canalJSONMessage2RowChange() *commonEvent.DML
 	columns := tableInfo.GetColumns()
 	switch msg.eventType() {
 	case canal.EventType_DELETE:
-		data := msg.getData()
-		appendRow2Chunk(data, columns, chk)
+		data := formatAllColumnsValue(msg.getData(), columns)
+		common.AppendRow2Chunk(data, columns, chk)
 	case canal.EventType_INSERT:
-		data := msg.getData()
-		appendRow2Chunk(data, columns, chk)
+		data := formatAllColumnsValue(msg.getData(), columns)
+		common.AppendRow2Chunk(data, columns, chk)
 	case canal.EventType_UPDATE:
-		previous := msg.getOld()
-		data := msg.getData()
+		previous := formatAllColumnsValue(msg.getOld(), columns)
+		data := formatAllColumnsValue(msg.getData(), columns)
 		for k, v := range data {
 			if _, ok := previous[k]; !ok {
 				previous[k] = v
 			}
 		}
-		appendRow2Chunk(previous, columns, chk)
-		appendRow2Chunk(data, columns, chk)
+		common.AppendRow2Chunk(previous, columns, chk)
+		common.AppendRow2Chunk(data, columns, chk)
 	default:
 		log.Panic("unknown event type for the DML event", zap.Any("eventType", msg.eventType()))
 	}
